feat: reorganize transport packages and update related tests; remove obsolete cache tests
