+++ CONTRIBUTING.md
@@ -33,13 +33,12 @@ packages/
   identity/            JWT verification, JWKS, scope, claims (TypeScript, Go, Python)
   revocation/          Revocation lookup interface and in-memory default (TypeScript, Go, Python)
   oauth/               RFC 8693 token exchange client (TypeScript)
-  transport-mcp/       MCP authentication core (TypeScript, Go, Python)
-  transport-a2a/       Agent-to-agent transport primitives (TypeScript)
+  transport/
+    mcp/               MCP authentication core (TypeScript, Go, Python)
+    a2a/               Agent-to-agent transport primitives (TypeScript)
   sdk/                 Caracal identity/delegation SDK (TypeScript, Python, Go)
   admin/               Typed admin API client (TypeScript)
-  framework-adaptor/   Per-framework adaptors over transport-mcp / sdk
-  runtime-adaptor/     Per-runtime adaptors (e.g. Cloudflare)
-  state-backend/       Storage backends (token state, token cache)
+  connectors/          Framework, runtime, and storage adapters (single-word dirs)
 infra/
   docker/              Compose orchestration and .env template
   postgres/migrations/ SQL migrations applied by the API at boot
