+++ go.work
@@ -1,15 +1,14 @@
 go 1.26
 
 use (
-	./services/sts
-	./services/audit
-	./services/gateway
-	./apps/coordinator/relay
-	./packages/core/go
-	./packages/identity/go
-	./packages/revocation/go
-	./packages/transport-mcp/go
-	./packages/caracalai-mcp/go
-	./packages/framework-adaptor/mcp-nethttp/go
-	./packages/sdk/go
+./services/sts
+./services/audit
+./services/gateway
+./apps/coordinator/relay
+./packages/core/go
+./packages/identity/go
+./packages/revocation/go
+./packages/transport/mcp/go
+./packages/connectors/nethttp/go
+./packages/sdk/go
 )
