+++ package.json
@@ -7,14 +7,14 @@
   },
   "scripts": {
     "caracal": "node apps/cli/bin/caracal-workspace.mjs",
-    "build:typescript": "pnpm --dir packages/core/ts build && pnpm --dir packages/oauth/ts build && pnpm --dir packages/admin/ts build && pnpm --dir packages/transport-a2a/ts build && pnpm --dir packages/identity/ts build && pnpm --dir packages/revocation/ts build && pnpm --dir packages/transport-mcp/ts build && pnpm --dir packages/framework-adaptor/mcp-express/ts build && pnpm --dir packages/framework-adaptor/mcp-fastmcp/ts build && pnpm --dir packages/framework-adaptor/agent-crewai/ts build && pnpm --dir packages/framework-adaptor/agent-langchain/ts build && pnpm --dir packages/state-backend/tokenstate-postgres/ts build && pnpm --dir packages/state-backend/tokencache-cloudflare/ts build && pnpm --dir packages/runtime-adaptor/cloudflare/ts build && pnpm --dir apps/api build && pnpm --dir apps/coordinator build",
+    "build:typescript": "pnpm --dir packages/core/ts build && pnpm --dir packages/oauth/ts build && pnpm --dir packages/admin/ts build && pnpm --dir packages/transport/a2a/ts build && pnpm --dir packages/identity/ts build && pnpm --dir packages/revocation/ts build && pnpm --dir packages/transport/mcp/ts build && pnpm --dir packages/connectors/express/ts build && pnpm --dir packages/connectors/fastmcp/ts build && pnpm --dir packages/connectors/postgres/ts build && pnpm --dir apps/api build && pnpm --dir apps/coordinator build",
     "lint": "pnpm -r --if-present lint",
     "typecheck": "pnpm -r --if-present typecheck",
     "ci": "pnpm --dir apps/cli sync-embedded && pnpm run build:typescript && pnpm run lint && pnpm run typecheck && pnpm run test",
     "test": "pnpm run test:typescript && pnpm run test:go && pnpm run test:python",
-    "test:go": "go test ./packages/core/go/... ./services/sts/... ./services/audit/... ./services/gateway/... ./apps/coordinator/relay/... ./packages/transport-mcp/go/... ./packages/framework-adaptor/mcp-nethttp/go/...",
-    "test:python": "PYTHONPATH=packages/identity/python:packages/revocation/python:packages/transport-mcp/python:packages/framework-adaptor/mcp-fastmcp/python python -m unittest discover -s tests/python -p 'test_*.py'",
-    "test:typescript": "pnpm --dir apps/api test && pnpm --dir apps/coordinator test && pnpm --dir apps/cli test && pnpm --dir apps/tui test && pnpm --dir packages/core/ts test && pnpm --dir packages/identity/ts test && pnpm --dir packages/admin/ts test && pnpm --dir packages/transport-a2a/ts test && pnpm --dir packages/oauth/ts test && pnpm --dir packages/framework-adaptor/mcp-express/ts test && pnpm --dir packages/framework-adaptor/agent-crewai/ts test && pnpm --dir packages/framework-adaptor/agent-langchain/ts test && pnpm --dir packages/state-backend/tokencache-cloudflare/ts test && pnpm --dir packages/runtime-adaptor/cloudflare/ts test"
+    "test:go": "go test ./packages/core/go/... ./services/sts/... ./services/audit/... ./services/gateway/... ./apps/coordinator/relay/... ./packages/transport/mcp/go/... ./packages/connectors/nethttp/go/...",
+    "test:python": "PYTHONPATH=packages/identity/python:packages/revocation/python:packages/transport/mcp/python:packages/connectors/fastmcp/python python -m unittest discover -s tests/python -p 'test_*.py'",
+    "test:typescript": "pnpm --dir apps/api test && pnpm --dir apps/coordinator test && pnpm --dir apps/cli test && pnpm --dir apps/tui test && pnpm --dir packages/core/ts test && pnpm --dir packages/identity/ts test && pnpm --dir packages/admin/ts test && pnpm --dir packages/transport/a2a/ts test && pnpm --dir packages/oauth/ts test && pnpm --dir packages/connectors/express/ts test"
   },
   "devDependencies": {
     "@vitest/coverage-v8": "3.2.4",
